Moved from tags to commit hashes
This is an experiment to see if commit hashes will work for our CI jobs.
The issue is we got affected by a bug in an github action which was
tagged. Therefor, the CI did not run. Also, it allows attackers to
inject malware into the action and moving the tag, so actions will be
automatically affected (see tj-action/changed-files).

By locking on specific versions will make sure always that particular
version will be used. However, this also means that we don't receive
security updates or bugfixes on that action.

Therefor, these are placed in dependabot which updates the actions on a
weekly interval.